Currently Added DSA

+++ Semester 1/Introduction to C/Week 5/Module 18.5/L_Summation.c
@@ -1,9 +1,9 @@
 // Starting in the name of Almighty Allah
 #include <stdio.h>
 
-int sum (int a[], int n, int i)
+long long int sum (long long int a[], long long int n, long long int i)
 {
-    int s = 0;
+    long long int s = 0;
     if (i == n) return 0;
     
     s = a[i] + sum (a, n, i+1);
@@ -12,15 +12,15 @@ int sum (int a[], int n, int i)
 
 int main ()
 {
-    int n;
+    long long int n;
     scanf("%d", &n);
     
-    int a[n];
+    long long int a[n];
     for (int i = 0; i < n; i++)
     {
         scanf("%d", &a[i]);
     }
-    int s = sum (a, n, 0);
+    long long int s = sum (a, n, 0);
     printf("%d", s);
     return 0;
 }
